Implement basic tokenizer + smaller changes.
Preliminary tokenization simplifies processing. strtok usually performs
additional operations not necessary for the purposes of this
program (e.g. checking against multiple delimiter characters).

* Eliminate i_basename:

Since the number of path components (depth) is known, it is unneccessary to save
the location of the basename.

* Rename i_midpath->subpath:

The new name communicates the content much more clearly. Additionally, the i
prefix suggests iteration, which is misleading.

* Remove malloc for output string:

No new string needs to be allocated, because the compressed path cannot be
longer than the full path (cwd). Therefore the former can be written in place of
the latter.